Seperate generateOutputval func from Calculator component for modularity
